Created a new package for main ui, moved the corresponding fragments there. Created a new package for NFTS ui, created a fragment and a view model there. Added dependencies for navigation, created navigation file.
