+++ app/build.gradle
@@ -64,6 +64,12 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-viewmodel-savedstate:2.5.0"
     implementation "androidx.lifecycle:lifecycle-reactivestreams-ktx:2.5.0"
 
+    // Navigation Kotlin
+    def nav_version="2.5.0"
+    implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
+    implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
+
+
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
