+++ app/src/main/java/com/example/learningapp/ui/main/MainViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.learningapp.ui
+package com.example.learningapp.ui.main
 
 import com.example.learningapp.base.viewmodel.BaseViewModel
