+++ app/src/main/java/com/example/learningapp/ui/nfts/NFTSFragment.kt
@@ -1,0 +1,4 @@
+package com.example.learningapp.ui.nfts
+
+class NFTSFragment {
+}
