+++ app/src/main/java/com/example/learningapp/ui/nfts/NFTSViewModel.kt
@@ -1,0 +1,6 @@
+package com.example.learningapp.ui.nfts
+
+import androidx.lifecycle.ViewModel
+
+class NFTSViewModel : ViewModel() {
+}
