+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.example.learningapp.ui.MainViewModel" />
+            type="com.example.learningapp.ui.main.MainViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
